fixed TodoListItem component so that it contains all 3 children components, a checkbox, text, and a trashcan
