+++ src/Components/TodoListItemText/TodoListItemText.css
@@ -1,5 +1,4 @@
-.todo-list-item {
+.todo-list-item-text {
     color: gray;
     font-size: 17px;
 }
-
