+++ src/Components/TodoListItems/TodoListItem.css
@@ -1,0 +1,20 @@
+.todo-list-item {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 40px 40px 65px;
+}
+.todo-list-item-right {
+    display: flex;
+    justify-content: start;
+    align-items: center;
+}
+
+.checkbox {
+    padding-right: 20px;
+}
+
+.trashcan {
+    color: gray;
+    font-size: 20px;
+}
